my-properties component integration with updated schema

+++ backend/src/main/java/com/stayhive/model/property/CancellationPolicy.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,11 +13,11 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CancellationPolicy {
-    private PolicyType type;          // Core tag group matching (FLEXIBLE, MODERATE, STRICT, CUSTOM)
-    private String name;              // Front UI Title text (e.g., "Moderate Tiered Policy")
-    private String description;       // Explanatory text for consumers
+    private PolicyType type;
+    private String name;
+    private String description;
 
-    // Supports multi-tier sequential timeline windows (ordered from highest daysBeforeCheckIn to lowest)
+
     private List<CancellationWindow> windows;
 
     @Data
@@ -24,8 +25,8 @@ public class CancellationPolicy {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class CancellationWindow {
-        private Integer daysBeforeCheckIn; // The day margin threshold required to clear (e.g., 14, 7, 0)
-        private Double refundPercentage;   // Total calculated return ratio window (e.g., 100.0, 50.0, 0.0)
+        private Integer daysBeforeCheckIn;
+        private BigDecimal refundPercentage;
 
     }
